Refactoring E2E
- Moving code around to encapsulate our e2e code under testutil.
- Creating multiple different go files that contain our various
e2e tests.
- Updated Makefile to allow for multiple tests.
- Created new tests to run with GitHub CI.
