+++ .github/workflows/ci.yaml
@@ -31,7 +31,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  e2e-kind:
+  e2e-kind-upgrades:
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -46,7 +46,41 @@ jobs:
           version: ${{ matrix.bazel }}
 
       - name: End-to-end (kind)
-        run: make test/e2e/kind
+        run: make test/e2e/kind-upgrades
+
+  e2e-kind-create:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        bazel: ["4.0.0"]
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Setup Bazel
+        uses: abhinavsingh/setup-bazel@v3
+        with:
+          version: ${{ matrix.bazel }}
+
+      - name: End-to-end (kind)
+        run: make test/e2e/kind-create
+
+  e2e-kind-decomission:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        bazel: ["4.0.0"]
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Setup Bazel
+        uses: abhinavsingh/setup-bazel@v3
+        with:
+          version: ${{ matrix.bazel }}
+
+      - name: End-to-end (kind)
+        run: make test/e2e/kind-decomission
 
   tests:
     runs-on: ubuntu-latest
@@ -54,7 +88,6 @@ jobs:
       fail-fast: false
       matrix:
         bazel: ["4.0.0"]
-
     steps:
       - uses: actions/checkout@v2
 
@@ -64,5 +97,4 @@ jobs:
           version: ${{ matrix.bazel }}
 
       - name: Tests
-        run: make test/all
-
+        run: make test/all
