+++ BUILD.bazel
@@ -63,7 +63,10 @@ filegroup(
         "//config:all-srcs",
         "//deploy/certified-metadata-bundle/cockroach-operator/latest/manifests:all-srcs",
         "//deploy/certified-metadata-bundle/cockroach-operator/latest/metadata:all-srcs",
-        "//e2e:all-srcs",
+        "//e2e/create:all-srcs",
+        "//e2e/decomission:all-srcs",
+        "//e2e/pvcresize:all-srcs",
+        "//e2e/upgrades:all-srcs",
         "//examples:all-srcs",
         "//hack:all-srcs",
         "//manifests:all-srcs",
