+++ e2e/create/BUILD.bazel
@@ -1,0 +1,43 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "go_default_test",
+    size = "enormous",
+    srcs = ["create_test.go"],
+    data = [
+        "//config/crd/bases:crd_manifest",
+        "//hack/bin:cockroach",
+        "//hack/bin:kind",
+        "//hack/bin:etcd",
+        "//hack/bin:kube-apiserver",
+        "//hack/bin:kubectl",
+        "//hack/bin:kubetest2",
+        "//hack/bin:kubetest2-kind",
+    ] + glob(["testdata/**"]),
+    deps = [
+        "//pkg/actor:go_default_library",
+        "//pkg/controller:go_default_library",
+        "//pkg/testutil:go_default_library",
+        "//pkg/testutil/env:go_default_library",
+        "@com_github_go_logr_zapr//:go_default_library",
+        "@com_github_stretchr_testify//require:go_default_library",
+        "@org_uber_go_zap//zaptest:go_default_library",
+    ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [
+        ":package-srcs",
+        "//e2e/kubetest2-openshift:all-srcs",
+    ],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
